Set options on new sockets in systematic manner
This patch consolidates the up-to-now scattered code that
sets different options on newly created sockets.

There are open_socket and open_tcp_socket functions that
do the tuning automatically.

In case the socket is not created but got from elsewhere
(such as accept() call) there are tune_socket and tune_tcp_socket
functions that will do the tuning.

Signed-off-by: Martin Sustrik <sustrik@250bpm.com>