Define _GNU_SOURCE in addition to the __BSD_VISIBLE on freebsd-like hosts
Define _GNU_SOURCE in addition to the __BSD_VISIBLE on the freebsd-like
hosts.  It made in hope to cover cases and obtain the full power of
underlying kernel on "hybrid" systems like Debian GNU/kFreeBSD, when
GLibc (or derived) library lives on top of FreeBSD kernel.

On the "real" FreeBSD it only enables getline(3), and therefore, should
not hurt.

* configure.ac: Define _GNU_SOURCE in addition to the __BSD_VISIBLE for
cover GLibc (or derived) on top of FreeBSD kernel cases, as it is in
Debian GNU/kFreeBSD case.