Improve efficiency of time measurement in tight loops
This patch instantiate a clock_t instance for each XS
socket. Thus, it is shared between subsequent calls
to xs_recv (and xs_send). That in turn significantly
limits the number of invocations of  getimeofday (or similar)
when timeouts are used and recv/send is called in a
tight loop.

Signed-off-by: Martin Sustrik <sustrik@250bpm.com>