Enabled plug-in loading on all hosts with support for dlopen()
Previously the code used for loading plug-ins would be built only on
Linux hosts. Now the required functions are checked for and plug-in
loading code is enabled if those are present on all platforms. Using
lstat() to check for file types also increases portability to hosts that
do not have the dirent.d_type field. If shared libraries support is
disabled plug-ins will also be automatically disabled.

Signed-off-by: Gabriele Svelto <gabriele.svelto@gmail.com>